Move Elasticsearch client code to its own package
